[SHAFAT]

+++ Bridgelabz/src/main/java/com/bridgelabz/service/FuncInterface.java
@@ -9,7 +9,7 @@ public interface FuncInterface {
 	public void buy(int amount,String symbol,int price);
 	public void sell(int amount,String symbol,int price);
 	public void save(String filename);
-	public double printReport();
+	public int printReport();
